requirements.txt
Django==4.0.1
django-grpc==1.0.21
django-oauth-toolkit==2.2.0
django-cors-headers==4.3.1
djangogrpcframework==0.2.1
djangorestframework==3.14.0
grpcio==1.51.3
grpc-django==1.0.0
grpcio-tools==1.51.3
oauthlib==3.2.2
psycopg2-binary==2.9.9